iis compiles again

+++ include/rikitiki/http/http_statuses.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <string>
 #include <codecvt>
+#include <stdint.h>
 
 namespace rikitiki {
   /**
@@ -10,12 +11,12 @@ namespace rikitiki {
    */
   class HttpStatus { 
   public:
-  HttpStatus(int s, const std::string& n, bool _managed = false) : status(s), name(n), managed(_managed) {}   
-  HttpStatus(int s, const std::wstring& n, bool _managed = false) : status(s), managed(_managed) {
+       HttpStatus(uint16_t s, const std::string& n, bool _managed = false) : status(s), name(n), managed(_managed) {}
+       HttpStatus(uint16_t s, const std::wstring& n, bool _managed = false) : status(s), managed(_managed) {
        std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>, wchar_t> conversion;
        name = conversion.to_bytes(n);
   }
-    int status; 
+    uint16_t status; 
     std::string name;
     bool managed; 
     std::string StartString() const;
